refactor(dir): ディレクトリ整理
